format: adjust format

+++ ffplayer/base/src/utility.cc
@@ -19,7 +19,7 @@ void update_thread_name(const char *name) {
 #if __linux__ || __ANDROID__
   pthread_setname_np(pthread_self(), name);
 #elif WIN32
-  int ret = MultiByteToWideChar(CP_UTF8, 0, name, strlen(name), nullptr, 0);
+  int ret = MultiByteToWideChar(CP_UTF8, 0, name, (int) strlen(name), nullptr, 0);
   if (ret < 0) {
     return;
   }
